Create a Clone operator, that will deal with both the local and remote operators
