+++ src/test/kotlin/com/github/jensim/megamanipulator/actions/localrepo/LocalRepoOperatorTest.kt
@@ -3,6 +3,7 @@ package com.github.jensim.megamanipulator.actions.localrepo
 import com.github.jensim.megamanipulator.actions.ProcessOperator
 import com.github.jensim.megamanipulator.actions.ProcessOperatorImpl
 import com.github.jensim.megamanipulator.actions.apply.ApplyOutput
+import com.github.jensim.megamanipulator.actions.git.localrepo.LocalRepoOperator
 import com.github.jensim.megamanipulator.ui.TestUiProtector
 import com.intellij.openapi.project.Project
 import io.mockk.every
