+++ src/test/kotlin/com/github/jensim/megamanipulator/actions/vcs/bitbucketserver/BitbucketServerClientTest.kt
@@ -1,7 +1,7 @@
 package com.github.jensim.megamanipulator.actions.vcs.bitbucketserver
 
 import com.github.jensim.megamanipulator.actions.NotificationsOperator
-import com.github.jensim.megamanipulator.actions.localrepo.LocalRepoOperator
+import com.github.jensim.megamanipulator.actions.git.localrepo.LocalRepoOperator
 import com.github.jensim.megamanipulator.http.HttpClientProvider
 import com.github.jensim.megamanipulator.settings.SettingsFileOperator
 import com.github.jensim.megamanipulator.settings.types.CloneType.HTTPS
